etl-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pandas==2.1.4
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
rq==1.15.1
